refactor(G1): split resolver/mod.rs into focused sub-modules (#87)
* refactor(G1): split resolver/mod.rs into fd, import_edit, hierarchy sub-modules

- resolver/fd.rs: package_prefix, import_file_stems, fd_find_and_parse,
  fd_search_by_full_path_pattern, fd_search_file, parse_fd_hits
- resolver/import_edit.rs: already_imported, import_insertion_line, make_import_edit
- resolver/hierarchy.rs: walk_hierarchy, HierarchyWalker, supertype_targets,
  super_names_for_class

mod.rs keeps the resolution chain (~520 lines, down from 1095).
External consumers unchanged — re-exports preserve public API paths.
830 tests pass, 0 new clippy warnings.


* refactor(G1): extract resolution chain into resolver/resolve.rs

resolver/mod.rs is now a 48-line thin coordinator (module declarations
+ re-exports only) — matching the rust-analyzer lsp.rs/cli.rs pattern.

The full resolution chain (ensure_file_data, fqns_for_name,
resolve_symbol, resolve_symbol_inner, resolve_symbol_no_rg, all step
implementations, rg_in_package_dir, is_stdlib, impl Indexer wrappers)
moves to resolver/resolve.rs (~700 lines, single responsibility).

No behaviour changes. 830 tests pass, 0 new clippy warnings.


* refactor(G1): address PR comments

- Update module docs: 'Kotlin and Java' → 'Kotlin, Java, and Swift'
- Replace #[allow(dead_code)] impl Indexer block with ResolutionChain trait:
  - Five resolution steps (local, via_imports, same_package, star_imports,
    qualified) become trait methods on ResolutionChain
  - impl ResolutionChain for Indexer delegates to the existing free functions
  - Removes _w naming suffix; no #[allow(dead_code)] on the impl block
  - TODO(G4) note: make resolve_symbol_inner generic over the trait
- Remove dead resolve_symbol_inner/fqns_for_name method wrappers from
  impl Indexer (callers already use the re-exported free functions)

830 tests pass, 0 new clippy warnings.


---------
